+++ yukawas/profiling_generate_and_train_all_nnsHOLO.py
@@ -1,4 +1,4 @@
-from cymetric.config import float_dtype, complex_dtype
+from cymetric.config import real_dtype, complex_dtype
 from cymetric.models.fubinistudy import FSModel
 import tensorflow as tf
 import tensorflow.keras as tfk
@@ -212,7 +212,7 @@ def train_and_save_nn(free_coefficient,nlayer=3,nHidden=128,nEpochs=50,stddev=0.
    phimodelzero.compile(custom_metrics=cmetrics)
 
    ## compare validation loss before training for zero network and nonzero network
-   datacasted=[tf.cast(data['X_val'],float_dtype),tf.cast(data['y_val'],float_dtype)]
+   datacasted=[tf.cast(data['X_val'],real_dtype),tf.cast(data['y_val'],real_dtype)]
    #need to re-enable learning, in case there's been a problem:
    phimodel.learn_transition = False
    phimodelzero.learn_transition = False
@@ -255,7 +255,7 @@ def train_and_save_nn(free_coefficient,nlayer=3,nHidden=128,nEpochs=50,stddev=0.
    print("ratio of final to zero: " + str({key + " ratio": value/(valzero[key]+1e-8) for key, value in valfinal.items()}))
    print("ratio of final to raw: " + str({key + " ratio": value/(valraw[key]+1e-8) for key, value in valfinal.items()}))
 
-   averagediscrepancyinstdevs,_=compute_transition_pointwise_measure(phimodel,tf.cast(data["X_val"],float_dtype))
+   averagediscrepancyinstdevs,_=compute_transition_pointwise_measure(phimodel,tf.cast(data["X_val"],real_dtype))
    print("average transition discrepancy in standard deviations: " + str(averagediscrepancyinstdevs))
    #IMPLEMENT THE FOLLOWING
    #meanfailuretosolveequation,_,_=measure_laplacian_failure(phimodel,data)
@@ -297,7 +297,7 @@ def load_nn_phimodel(free_coefficient,nlayer=3,nHidden=128,nEpochs=50,bSizes=[19
    #nn_phi_zero = make_nn(n_in,n_out,nlayer,nHidden,act,use_zero_network=True)
    print("nns made")
 
-   datacasted=[tf.cast(data['X_val'],float_dtype),tf.cast(data['y_val'],float_dtype)]
+   datacasted=[tf.cast(data['X_val'],real_dtype),tf.cast(data['y_val'],real_dtype)]
 
    #    nn_phi = make_nn(n_in,n_out,nlayer,nHidden,act,use_zero_network=True)
    #    nn_phi_zero = make_nn(n_in,n_out,nlayer,nHidden,act,use_zero_network=True)
@@ -380,7 +380,7 @@ def load_nn_phimodel(free_coefficient,nlayer=3,nHidden=128,nEpochs=50,bSizes=[19
    print("validation loss for final network: ")
    print(valtrained)
    print("ratio of trained to zero: " + str({key + " ratio": value/(valzero[key]+1e-8) for key, value in valtrained.items()}))
-   averagediscrepancyinstdevs,_=compute_transition_pointwise_measure(phimodel,tf.cast(data["X_val"],float_dtype))
+   averagediscrepancyinstdevs,_=compute_transition_pointwise_measure(phimodel,tf.cast(data["X_val"],real_dtype))
    print("average transition discrepancy in standard deviations: " + str(averagediscrepancyinstdevs))
    print("\n\n")
    #IMPLEMENT THE FOLLOWING
@@ -452,7 +452,7 @@ def train_and_save_nn_HYM(free_coefficient,linebundleforHYM,nlayer=3,nHidden=128
    databeta_val=tf.data.Dataset.from_tensor_slices(databeta_val_dict)
    # batch_sizes=[64,10000]
    databeta_train=databeta_train.shuffle(buffer_size=1024).batch(bSizes[0],drop_remainder=True)
-   datacasted=[tf.cast(databeta['X_val'],float_dtype),tf.cast(databeta['y_val'],float_dtype)]
+   datacasted=[tf.cast(databeta['X_val'],real_dtype),tf.cast(databeta['y_val'],real_dtype)]
 
    cb_list, cmetrics = getcallbacksandmetricsHYM(databeta)
 
@@ -517,7 +517,7 @@ def train_and_save_nn_HYM(free_coefficient,linebundleforHYM,nlayer=3,nHidden=128
    betamodel.compile(custom_metrics=cmetrics)
    betamodelzero.compile(custom_metrics=cmetrics)
    
-   #datacasted=[tf.cast(data['X_val'],float_dtype),tf.cast(data['y_val'],float_dtype)]
+   #datacasted=[tf.cast(data['X_val'],real_dtype),tf.cast(data['y_val'],real_dtype)]
    print("testing zero and raw")
    valzero=betamodelzero.test_step(databeta_val_dict)
    valraw=betamodel.test_step(databeta_val_dict)
@@ -587,7 +587,7 @@ def train_and_save_nn_HYM(free_coefficient,linebundleforHYM,nlayer=3,nHidden=128
    print("ratio of final to raw: " + str({key + " ratio": value/(valraw[key]+1e-8) for key, value in valfinal.items()}))
 
 
-   averagediscrepancyinstdevs,_=compute_transition_pointwise_measure(betamodel,tf.cast(databeta["X_val"],float_dtype))
+   averagediscrepancyinstdevs,_=compute_transition_pointwise_measure(betamodel,tf.cast(databeta["X_val"],real_dtype))
    print("average transition discrepancy in standard deviations: " + str(averagediscrepancyinstdevs))
 
 
@@ -692,7 +692,7 @@ def load_nn_HYM(free_coefficient,linebundleforHYM,nlayer=3,nHidden=128,nEpochs=3
    print("ratio of trained to zero: " + str({key + " ratio": value/(valzero[key]+1e-8) for key, value in valtrained.items()}))
 
 
-   averagediscrepancyinstdevs,_=compute_transition_pointwise_measure(betamodel,tf.cast(databeta["X_val"],float_dtype))
+   averagediscrepancyinstdevs,_=compute_transition_pointwise_measure(betamodel,tf.cast(databeta["X_val"],real_dtype))
    print("average transition discrepancy in standard deviations: " + str(averagediscrepancyinstdevs))
    import time
    start=time.time()
@@ -978,7 +978,7 @@ def train_and_save_nn_HF(free_coefficient,linebundleforHYM,betamodel,functionfor
    #print("perm10")
    #print(perm.print_diff())
 
-   averagediscrepancyinstdevs,_=compute_transition_pointwise_measure_section(HFmodel,tf.cast(dataHF["X_val"],float_dtype))
+   averagediscrepancyinstdevs,_=compute_transition_pointwise_measure_section(HFmodel,tf.cast(dataHF["X_val"],real_dtype))
    print("average transition discrepancy in standard deviations (note, underestimate as our std.dev. ignores variation in phase): " + str(averagediscrepancyinstdevs))
 
 
@@ -1093,7 +1093,7 @@ def load_nn_HF(free_coefficient,linebundleforHYM,betamodel,functionforbaseharmon
    print("ratio of trained to zero: " + str({key + " ratio": value/(valzero[key]+1e-8) for key, value in valtrained.items()}))
 
 
-   averagediscrepancyinstdevs,_=compute_transition_pointwise_measure_section(HFmodel,tf.cast(dataHF["X_val"],float_dtype))
+   averagediscrepancyinstdevs,_=compute_transition_pointwise_measure_section(HFmodel,tf.cast(dataHF["X_val"],real_dtype))
    print("average transition discrepancy in standard deviations (note, underestimate as our std.dev. ignores variation in phase): " + str(averagediscrepancyinstdevs))
    meanfailuretosolveequation,_,_=HYM_measure_val_with_H(HFmodel,dataHF)
    print("mean of difference/mean of absolute value of source, weighted by sqrt(g): " + str(meanfailuretosolveequation))
